+++ CHANGELOG.md
@@ -42,9 +42,15 @@ Bottom level categories:
 
 For naga changelogs at or before v0.14.0. See [naga's changelog](naga/CHANGELOG.md).
 
-### Bug fixes
+### Bug Fixes
 
-- Fix issue in the metal backend of naga, where local variables were sometimes using variable names from previous functions
+#### WGL
+
+- Create a hidden window per `wgpu::Instance` instead of sharing a global one.
+
+#### Naga MSL-OUT
+
+- Fix issue where local variables were sometimes using variable names from previous functions.
 
 ## v0.18.0 (2023-10-25)